parameter optimization code, not finished yet
neural networks with keras library to optimize parameters and build an optimized risk diagram. the first functions are for data analysis, towards the end of the notebook i began building a neural net. so far, it calculates okay the rt (not averaged yet) values. checked them, they are correct values.