optimization idea 2 works!!!
the loss doesn't crush, found the non-diff part of it all, limiting the data, it's awesome! it's logically very fast bc GPU, and could train for epochs on end. 
may need cross-validation and etc, but so far so good with it working! idk how it'll look with the predictions, but from other stuff i've tried i feel the changes will be significant. it's a global approach, taking all data.